Optimizing (really making not terrible):
Instead of getting all reads for a group, get an iterator of reads, and
only keep informative reads to reduce memory footprint.  Write those
reads to output BAM immediately and discard

Don't call removeAll on 2 lists.  That hurts..bad.

Extra logging for the number of reads kept in memory and number of tags
collapsed.