Fix for an edge case where reads disagree for a UMI (#200)
Very rare edge case where a UMI is observed with exactly 2 reads
with two different bases.  They disagree, so the "most common" base is
randomly one of the two.  This could cause the collection of bases to
have an average error rate close to 1, and Phred encoding would round it
to 0.  This would then cause some likelihood scores to be log (0), which
is bad m'kay?