Completed modifications to scRNASeq donor assignment code.
For a given UMI, only the best (highest average genotype quality in the
population) SNP is tagged for the read group.  This means that for a
given UMI, only a single SNP is counted as informative.

This removes some "double counting" that was occurring previously, where
multiple SNPs for a single UMI could be counted as independent
observations.