FiltererIterator modified to track reads that pass/fail
FiltererIterator now tracks reads that pass/fail filterOut
automatically.  When the iterator is exhausted, a call to an abstract
method logFilterResults is made (which is implemented by the subclass)
that logs the state of the iterator.

All FiltererIterator implementations were updated to implement the new
logging method, though most were very boilerplate.

Added some quality of life checks to VCF donor list checks to report if
there were duplicate donors in the input list, or if donors were
requested that were missing from the VCF the identities of the missing
donors is reported, which should aid users fixing the problem.

Added a way for SNPUMICellReadIteratorWrapper2 to fail quickly if it
sees many UMIs that do not contain transcribed SNPs.  This part needs
code review - an unchecked exception is thrown when this fails which may
not be optimal, and the implementation might be improved.