Jn dropulation fix doublecount2 (#344)
* Refactored DNA based SNP pileups to only use a single SNP per read.  If
there are multiple SNPs overlapping a read, the best SNP (with the
highest GQ score) is selected for that read, and all other SNPs are
ignored.

* Modified Test data and unit tests for refactoring / SNP Pileup work

* Test data and new unit tests

* SNPUMICellReadIteratorWrapper should always be using meanGenotypeQuality

* implementation notes

* ...

* ...

* Cleaned up unused progress logger

* A new version of SNPUMICellReadIteratorWrapper that only tags a single
SNP for a group of reads that is defined by a UMI (gene/cell/UMI
sequence)

* Fixing java warnings: close writer after use.

* cleanup imports

* Fixed java warning about non-static reference to variable.

* Completed modifications to scRNASeq donor assignment code.

For a given UMI, only the best (highest average genotype quality in the
population) SNP is tagged for the read group.  This means that for a
given UMI, only a single SNP is counted as informative.

This removes some "double counting" that was occurring previously, where
multiple SNPs for a single UMI could be counted as independent
observations.

* Slight modification for debugging.

* FiltererIterator modified to track reads that pass/fail

FiltererIterator now tracks reads that pass/fail filterOut
automatically.  When the iterator is exhausted, a call to an abstract
method logFilterResults is made (which is implemented by the subclass)
that logs the state of the iterator.

All FiltererIterator implementations were updated to implement the new
logging method, though most were very boilerplate.

Added some quality of life checks to VCF donor list checks to report if
there were duplicate donors in the input list, or if donors were
requested that were missing from the VCF the identities of the missing
donors is reported, which should aid users fixing the problem.

Added a way for SNPUMICellReadIteratorWrapper2 to fail quickly if it
sees many UMIs that do not contain transcribed SNPs.  This part needs
code review - an unchecked exception is thrown when this fails which may
not be optimal, and the implementation might be improved.

* Additional data for unit testing

* FilteredIterator::logFilterResults is now a default no-op.