Add command-line option PAIRED_MODE
PAIRED_MODE:  Require user to specify if reads should not be processed in pairs,
rather than old behavior in which an accidentally missing mate would cause a read to be processed as unpaired.

Also, implemented by not exposed:
ALLOW_UNSORTED: don't queryname sort input, which requires read 2 to follow read 1 even if not queryname sorted.
This option is prone to misuse and we don't have the use case we thought we had, so leave implemented but
don't expose.