Update to how we deal with alternate alleles especially at multiallelic sites and how alternate alleles are calculated from the VCF.  These are rare events and either don't change results at all, or have a minimal effect on unit tests (1 SNP in 1000 or fewer changes).
When analysis subset samples from a VCF site for analysis, it was recomputing the alleles.  This could change sites that had an alternate allele in the original VCF to be classified as NO_VARIATION in the new site object, which lost the alternate allele.  These alleles were then propagated as N in downstream analysis.  This behavior was changed to not recompute the alleles, which preserves the original ref/alt alleles in the VCF. This made it necessary to not rely on the NO_VARIATION tag to filter SNPs that are not polymorphc in the subset of samples being analyzed.  In most analysis, this has no effect as we only care about polymorphic sites in the population, but severely broke GatherDigitalAlleleCountsTest when POLYMORPHIC_SNPS_ONLY was set to false.

If a multiallelic site has been split into two biallelic sites, the alternate allele and site quality were associated to an interval (chromosome:start:end) of a site, which caused these features to be associated with the last SNP encountered in the VCF.

If both alleles were common, then the quality of one was arbitrarily selected, but both intervals could be  tested downstream, which could lead to rare double counting of some sites.

Instead, for all variants that share the same interval, select the pair variant with the greatest minor allele frequency that will be more informative for downstream analysis and ignore the other alleles to avoid any possibility of double counting.