Alignment script: Add `ncores` parameter (for STAR)
This adds an (optional) parameter `ncores` to the
`Drop-seq_alignment.sh` script, to enable running STAR with more than
just one thread.
This greatly improves the run-time of the script, if used.
The default value was chosen such that it matches STAR's default for the
`--runThreadN` parameter. Thus, the bahaviour of the script remains
identical when omitting the optional new parameter.