Fix indentation errors introduced by 957e982
Commit 957e98205b4126582457ad4aae68fada9ada2872 introduced the
`mark_file_as_intermediate` convenience-alias to `Drop-seq_alignment.sh`
to work around the lack of arrays in POSIX shells.
To reduce the risk of breaking anything due to this change, regular
expression replacement was used to adjust the code accordingly.
By mistake, this introduced a level of indentation to *all* calls of
that alias, not only the one instance wrapped into a conditional block.
While this is only a cosmetic issue that has no functional consequences,
correcting the indentation of the affected lines improves readability
and thus maintainability of the code.