Shell scripts: Remove empty assignments
`param=` for an uninitialised variable `param` results in `$param`
expanding to `""`, just as it would without.
While the idea of declaring initialising all parameters upfront is nice,
the current state of the code shows that the maintenance burden to keep
this consistent was too high in the past, suggesting it will be in the
future as well: Some parameters are initialised by setting them empty,
some are unconditionally unset, and some are not initialised at all.
The easiest course of action to standardize this seems to be removing
all no-op lines, which is (partially) done by this commit.