Drop empty variable expansion / superfluous space
The `filtered_gene_biotypes` parameter was initialized empty but
expanded upon its (conditional) assignment in a self-appending manner.
Effectively, this only added an extra leading space before the `G`
parameters being set in that line. Since the variable is expanded as
part of a command where it is preceded by a space, the only effect this
self-reference had, was to introduce an extra space that has no effect.
Removing this construct improves readability and thus maintainability of
the code.