Shell scripts: Return sucess (0) upon `-h`
If a script is called with the `-h` flag, the usage is printed and the
script exists. This is the expected behaviour. Thus, returning a
non-zero value (indicating an error) is confusing.

This would have made *e.g.* troubleshooting
https://github.com/broadinstitute/Drop-seq/issues/397#issuecomment-1999818084
easier, as all script come with a `-h` flag that could safely be called
without expecting anything to go wrong.