Shell scripts: Unset uninitialized parameters
Commit a1bc4797dda88121facda747e78f309e3fed7beb removed empty
(initialization) assignments and
3fc809ebd57ba4f73b13e3b9e8489f6131ee32f6 did the same for 'no-op'
unset operations.
The rationale was that those (empty) initializations would no have any
effect (in the context of the script itself) and while the idea of
initializing each and every parameter seemed nice, the state of the code
suggested the maintenance burden was too high.
However, in
https://github.com/broadinstitute/Drop-seq/pull/412#discussion_r1569231368
Alec Wysoker (@alecw) asked to re-instate these empty initializations
'in order to ensure they're passed on command line rather than inherited
from somewhere'.

Instead of simply reverting the removal of the previous empty
initializations, this commit systematically unsets all variables to be
set by command line parameters that have not been initialized before
just before parsing the command line.
Sourced scripts have been considered as well for potential
initializations and comments have been added to document this design
choice.
For consistency, the order of parameters within the (single) unset
statements was chosen to match the order of the parameters (which has
already been standardizes between the usage string and the command line
parsing `case` statements).

Note that capitalized variables have been excluded from this
unconditional empty initialization as an all-capital variable name
usually flags a variable as an environment variable that is meant to be
overridden outside of the script by the shell calling the script.