Alignment script: Refactor using aliases
This simply refactors the code by pre-defining commands for later use.
As commands can contain variable parameters that themselves might
contain spaces, this was implemented via aliases.
In Dash, this works out of the box. In Bash, `shopt expand_aliases` is
required. Thus, it is enabled, if necessary.

This is in preparation to introducing a pipeline mode that speeds up the
script by chaining commands via pipes to avoid I/O overhead caused by
compressing, writing, reading, and uncompressing intermediate files.